Reimplement some of the BitOps methods
They were buggy. Such is life :p

These tests cover the various boundary conditions int/uint/long/ulong
max and min values, as well as values which expand from 32bit -> 33bit
numbers. The latter covers cases like rounding up uint.MaxValue to the
next power of two.

+++ src/MonoTorrent.BEncoding/MonoTorrent.BEncoding/BEncodedNumber.cs
@@ -116,7 +116,7 @@ namespace MonoTorrent.BEncoding
                 case long.MinValue: // -9223372036854775808
                     return 2 + 20;
                 default:
-                    return 2 + 1 + (Number < 0 ? 1 : 0) + BitOps.CeilLog10 ((ulong) Math.Abs (Number));
+                    return 2 + 1 + (Number < 0 ? 1 : 0) + BitOps.CeilLog10 (Math.Abs (Number));
             }
         }
